# crates/mantissa-rest/openapi/mantissa-rest.openapi.json
{
  "components": {
    "schemas": {
...
      "AgentSessionSummary": {
        "description": "REST-facing compact agent session row.",
        "properties": {
          "active_run_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "execution_platform": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isolation_mode": {
            "type": "string"
          },
          "isolation_profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_run_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "execution_platform",
          "isolation_mode",
          "updated_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "Mantissa REST token",
        "description": "Daemon-generated REST token presented as a Bearer token.",
        "scheme": "bearer",
        "type": "http"
      },
      "mtls": {
        "description": "Required when the REST listener is bound to a non-loopback interface.",
        "type": "mutualTLS"
      }
    }
  },
  "info": {
    "description": "Local HTTP facade over the Mantissa Cap'n Proto admin session. All /v1 endpoints require an Authorization: Bearer token. Non-loopback listeners must also be served with TLS and client certificate verification enabled.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Mantissa REST API",
    "version": "0.1.0"
  },
  "openapi": "3.2.0",
  "paths": {
...
    "/v1/agents/sessions": {
      "get": {
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentSessionSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Durable agent sessions."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists durable agent sessions visible to the local daemon.",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "submit_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSubmitResponse"
                }
              }
            },
            "description": "Submitted agent session metadata."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Submits one durable agent session manifest to the local daemon.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSession"
                }
              }
            },
            "description": "Deleted agent session snapshot."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Deletes one closed durable agent session and its retained run history.",
        "tags": [
          "agents"
        ]
      },
      "get": {
        "operationId": "get_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSessionDetail"
                }
              }
            },
            "description": "Agent session detail."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Fetches one durable agent session and its retained run history.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}/cancel": {
      "post": {
        "operationId": "cancel_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSession"
                }
              }
            },
            "description": "Updated agent session snapshot."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Requests cancellation for one active or queued agent session run.",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/v1/clusters/current": {
      "get": {
        "operationId": "current",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterView"
                }
              }
            },
            "description": "Current active cluster view."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Returns the active cluster view associated with the local session.",
        "tags": [
          "clusters"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}